Remove all problems with unsigned/signed integers (hopefully correct), and in the process, cast some -1s to their right types...
darcs-hash:20080702161228-3bc9d-f829e006d52173111c320c7b1b90910d2a1a4634

+++ file.h
@@ -37,7 +37,7 @@ struct FdAndMap {
     int fd;
     char* map;
     char* preloaded;
-    size_t size;
+    off_t size; // ### for fstat...
     size_t preloadsize;
 };
 typedef enum { Create, Read, ReUse } FileOpenMode; // ReUse assumes file exists
